@@ -1,52 +1,61 @@
 {
   "name": "@primno/cli",
-  "version": "0.3.0",
-  "description": "Create, build and serve primno workspace",
-  "main": "dist/index.js",
+  "version": "0.6.0-beta.0",
+  "description": "Command-line interface tool for initializing, building, and deploying Primno projects",
+  "main": "dist/index.mjs",
   "files": [
     "dist"
   ],
   "repository": "github:primno/cli",
   "scripts": {
-    "start": "node dist/index.js",
+    "start": "rollup -c && node dist/index.mjs",
     "build": "rollup -c",
     "watch": "rollup -cw"
   },
   "bin": {
-    "mn": "dist/index.js"
+    "mn": "dist/index.mjs"
+  },
+  "engines": {
+    "node": ">=16.0.0"
   },
   "author": "Xavier Monin",
-  "license": "ISC",
+  "license": "MIT",
   "dependencies": {
-    "@primno/d365-client": "^0.6.1-beta.2",
-    "@rollup/plugin-commonjs": "^21.0.3",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.1.3",
-    "@rollup/plugin-typescript": "^8.3.1",
-    "@rollup/plugin-virtual": "^2.1.0",
-    "chalk": "^4.1.2",
+    "@primno/dataverse-client": "^0.8.0-beta.1",
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-terser": "^0.4.0",
+    "@rollup/plugin-typescript": "^11.0.0",
+    "@rollup/plugin-virtual": "^3.0.1",
+    "chalk": "^5.2.0",
     "commander": "^9.1.0",
     "cors": "^2.8.5",
-    "express": "^4.17.3",
+    "express": "^4.18.2",
     "figlet": "^1.5.2",
-    "glob": "^8.0.3",
-    "listr2": "^4.0.5",
+    "glob": "^8.1.0",
+    "listr2": "^5.0.7",
     "mustache": "^4.2.0",
-    "rollup": "^2.70.1",
-    "rollup-plugin-terser": "^7.0.2",
+    "rollup": "^3.15.0",
     "rxjs": "^7.5.5",
-    "selfsigned": "^2.0.1",
-    "terser": "^5.13.1",
-    "tslib": "^2.3.1",
-    "typescript": "^4.6.3"
+    "selfsigned": "^2.1.1",
+    "terser": "^5.16.3",
+    "tslib": "^2.5.0",
+    "typescript": "^4.9.5"
   },
   "devDependencies": {
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/figlet": "^1.5.4",
-    "@types/glob": "^7.2.0",
-    "@types/mustache": "^4.1.2",
-    "@types/node": "~16.10.3",
+    "@types/glob": "^8.0.1",
+    "@types/mustache": "^4.2.2",
+    "@types/node": "^18.13.0",
     "@types/rollup__plugin-virtual": "^2.0.1"
+  },
+  "peerDependencies": {
+    "@primno/core": ">= 0.6.0-beta.0 < 0.7.0-beta.0"
+  },
+  "publishConfig": {
+    "access": "public",
+    "registry": "https://registry.npmjs.org/"
   }
 }
