@@ -1,7 +1,7 @@
-import { nodeResolve } from '@rollup/plugin-node-resolve';
+import resolve from '@rollup/plugin-node-resolve';
 import typescript from "@rollup/plugin-typescript";
 import commonjs from '@rollup/plugin-commonjs';
-import pkg from './package.json';
+import pkg from './package.json' assert { type: "json" };
 
 const external = [
     ...Object.keys(pkg.dependencies || {}),
@@ -9,9 +9,9 @@ const external = [
 ];
 
 const plugins = [
-    nodeResolve(),
+    typescript(),
     commonjs(),
-    typescript({ module: "esnext" })
+    resolve()
 ];
 
 export default [
@@ -20,6 +20,6 @@ export default [
         input: 'src/index.ts',
         plugins,
         external,
-        output: { format: 'cjs', file: 'dist/tsc.js', sourcemap: "inline" },
+        output: { format: 'esm', file: 'dist/tsc.mjs', sourcemap: "inline" },
     }
 ];
