@@ -1,41 +1,45 @@
-import { D365Client } from "@primno/d365-client";
+import { DataverseClient } from "@primno/dataverse-client";
 import { SolutionRepository } from "../d365/repository/solution-repository";
 import { readFile } from "fs/promises";
 import { SolutionComponentType } from "../d365/model/add-solution-component";
-import { WebResourceType } from "../d365/model/webresource";
-import { WebResourceRepository } from "../d365/repository/webresource-repository";
+import { WebResourceType } from "../d365/model/web-resource";
+import { WebResourceRepository } from "../d365/repository/web-resource-repository";
 import { isNullOrUndefined } from "../../utils/common";
 import { Solution } from "../d365/model/solution";
-import { getCacheDir } from "../../utils/cache";
+import { defaultConnectionString, defaultSolutionUniqueName } from "../../configuration/workspace-configuration";
+import { DataverseOptions } from "./dataverse-options";
+import { getClient } from "../../utils/dataverse-client";
 
-export interface DeployerConfig {
-    // TODO: Change location ?
+export interface DeployerOptions extends DataverseOptions {
     projectName: string;
-    connectionString: string;
     solutionUniqueName: string;
+
+    /**
+     * Callback for device code authentication.
+     */
+    deviceCodeCallback: (url: string, code: string) => void;
 }
 
-export abstract class Deployer<TCfg extends DeployerConfig> {
-    public constructor(protected sourcePath: string, protected config: TCfg) { }
+export abstract class Deployer<TOptions extends DeployerOptions> {
+    public constructor(protected sourcePath: string, protected config: TOptions) {
+        if (config.environment?.connectionString == null ||
+            config.environment?.connectionString === defaultConnectionString) {
+            throw new Error("Invalid connection string");
+        }
+
+        if (config.solutionUniqueName == null || config.solutionUniqueName === defaultSolutionUniqueName) {
+            throw new Error("Invalid solution name");
+        }
+    }
 
     public async deploy(): Promise<string> {
         try {
-            const d365Client = new D365Client(
-                this.config.connectionString,
-                {
-                    persistence: {
-                        enabled: true,
-                        cacheDirectory: getCacheDir(),
-                        serviceName: "primno-cli",
-                        accountName: this.config.connectionString
-                    }
-                }
-            );
+            const client = getClient(this.config.environment.connectionString, this.config.deviceCodeCallback);
 
-            const solutionRepository = new SolutionRepository(d365Client);
-            const webResourceRepository = new WebResourceRepository(d365Client);
+            const solutionRepository = new SolutionRepository(client);
+            const webResourceRepository = new WebResourceRepository(client);
 
-            // Exist ?
+            // Exists ?
             const solution = await solutionRepository.getByName(this.config.solutionUniqueName);
 
             if (isNullOrUndefined(solution)) {
@@ -49,6 +53,8 @@ export abstract class Deployer<TCfg extends DeployerConfig> {
             const webResourceId = await webResourceRepository.createOrUpdate({
                 content: fileContent,
                 name: webResourceName,
+                displayname: `${this.config.projectName} - ${webResourceName}`,
+                description: `Web resource generated by Primno CLI`,
                 webresourcetype: WebResourceType.JS,
                 solutionid: solution.solutionid
             });
