@@ -1,109 +1,10 @@
 import { Listr, ListrTask } from "listr2";
 import { Observable } from "rxjs";
-
-type ActionHandler = () => string | void | Promise<string | void> | Observable<string>;
+import { ActionHandler, TaskInfo, TaskType } from "./task-info";
 
 export interface Action {
     title: string;
     action: ActionHandler;
-}
-
-enum TaskType {
-    /**
-     * Function to call
-     */
-    Action,
-    Observable,
-    /**
-     * Task containing subtasks.
-     */
-    Level,
-    /**
-     * Root task (level 0).
-     * This type of task has no real existence and will not be displayed.
-     * Used only to allow the creation of tasks when using {@link Task.new()}.
-     */
-    Root
-}
-
-class TaskInfo {
-    private _title?: string;
-    private _action?: ActionHandler;
-    private _concurrent?: number | boolean = false;
-    private _type: TaskType;
-    private _observable?: Observable<string>;
-    private _exitOnError: boolean = true;
-
-    private constructor(type: TaskType, title?: string) {
-        this._title = title;
-        this._type = type;
-    }
-
-    public static newAction(title: string, action: ActionHandler) {
-        const task = new TaskInfo(TaskType.Action, title);
-        task._action = action;
-
-        return task;
-    }
-
-    public static newObservable(title: string, observable: Observable<string>) {
-        const task = new TaskInfo(TaskType.Observable, title);
-        task._observable = observable;
-
-        return task;
-    }
-
-    public static newLevel(title?: string) {
-        return new TaskInfo(TaskType.Level, title);
-    }
-
-    public static newRoot() {
-        return new TaskInfo(TaskType.Root);
-    }
-
-    public get type(): TaskType  {
-        return this._type;
-    }
-
-    public set type(value) {
-        this._type = value;
-    }
-
-    public get title() {
-        return this._title;
-    }
-
-    public set title(value) {
-        this._title = value;
-    }
-
-    public get action() {
-        return this._action;
-    }
-
-    public get observable() {
-        return this._observable;
-    }
-
-    public get concurrent() {
-        return this._concurrent;
-    }
-
-    public set concurrent(value) {
-        this._concurrent = value;
-    }
-
-    public get exitOnError() {
-        return this._exitOnError;
-    }
-
-    public set exitOnError(value) {
-        this._exitOnError = value;
-    }
-
-    public get persistentOutput() {
-        return true;
-    }
 }
 
 export class Task {
@@ -188,7 +89,7 @@ export class Task {
     /**
      * Sets how many tasks will be run at the same time in parallel.
      * @param concurrent Concurrency of subtasks.
-     *  false: no concurrency, all tasks run synchronously. true: subtasks run simulatneously. number: limit the max concurrency.
+     *  false: no concurrency, all tasks run synchronously. true: subtasks run simultaneously. number: limit the max concurrency.
      * @returns The current task.
      */
     public withConcurrency(concurrent: number | boolean): Task {
@@ -216,7 +117,7 @@ export class Task {
      * Build ListR task.
      * @returns Built ListrTask
      */
-    private getListRTask(): ListrTask | null | ListrTask[] {
+    private getListRTask(): ListrTask | ListrTask[] | null  {
         switch (this.type) {
             case TaskType.Action:
                 return {
@@ -243,6 +144,9 @@ export class Task {
                             if (typeof result === "string" && !rootTask.hasFailed()) {
                                 t.output = result;
                             }
+                            else {
+                                return result;
+                            }
                         } catch(except) {
                             if (!rootTask.hasFailed()) {
                                 throw except;
@@ -256,7 +160,10 @@ export class Task {
             case TaskType.Observable:
                 return {
                     title: this.taskInfo.title,
-                    task: async (ctx, t) => this.taskInfo.observable
+                    task: (ctx, t) => this.taskInfo.observable,
+                    options: {
+                        persistentOutput: this.taskInfo.persistentOutput
+                    }
                 };
             case TaskType.Level:
             case TaskType.Root:
@@ -294,7 +201,7 @@ export class Task {
     private getListR(): Listr {
         const listRTask = this.getListRTask();
 
-        if (listRTask === null) {
+        if (listRTask == null) {
             throw new Error("An empty task can't be run");
         }
 
@@ -306,10 +213,11 @@ export class Task {
                 collapseErrors: false,
                 showErrorMessage: true,
                 showTimer: true,
-                collapse: false
+                collapse: false,
+                lazy: true // Remove the showing spinner
             },
             ...this.getListROptions(),
-            exitOnError: true,
+            // exitOnError: true
         });
 
         return listR;
